test/requirements.txt
gunicorn==19.9.0
gevent==1.3.5

django==1.11
djangorestframework==3.8.2
django-cached-authentication-middleware==0.2.2
django-cors-middleware==1.3.1
channels==1.1.8
asgi_redis==1.4.3
service-identity==17.0.0
Twisted[tls,http2]==18.7.0

mysqlclient==1.3.13
django-redis==4.10.0
python-memcached==1.59
requests==2.19.1
Pillow==5.2.0
paramiko==2.4.1
psutil==5.4.7
retry==0.9.2

python-openstackclient==3.14.1
python-novaclient==9.1.1
python-neutronclient==6.7.0
python-glanceclient==2.10.0
python-keystoneclient==3.15.0
python-cinderclient==3.5.0
python-ceilometerclient==2.9.0
python-zunclient==1.1.0
docker==3.4.1

pyminizip==0.2.4
thriftpy==0.3.9

mkdocs==0.17.2
mkdocs-material==2.7.3

protobuf==3.6.1

Cython==0.29.6
ConcurrentLogHandler==0.9.1